contactsapi/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.0
virtualenv-clone==0.5.6
